requirements.txt
# TalentFlow AI - Requirements
# Python 3.10+

# === Frontend (Streamlit) ===
streamlit==1.42.0
Authlib==1.3.2
audio-recorder-streamlit==0.0.10

# === Backend (FastAPI) ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# === Database ===
motor==3.7.0  # Async MongoDB driver
pymongo==4.10.1  # Keep original version - works fine for our use case
dnspython==2.7.0  # For MongoDB connection

# === AI/ML Libraries ===
# OpenAI & Azure
openai==1.59.5

# LangChain - Use compatible versions
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langchain-core==0.3.28  # Keep original - compatible with langchain 0.3.13

# LangGraph for AI workflows
langgraph==0.2.59
langgraph-checkpoint==2.0.10  # Keep original version

# === PDF Processing ===
PyPDF2==3.0.1
fpdf==1.7.2

# === Azure Services ===
azure-cognitiveservices-speech==1.41.1

# === Voice Processing (for AI Interview - Optional) ===
sounddevice==0.5.1
scipy==1.15.1
playsound==1.2.2
numpy==1.26.4  # Compatible with scipy and sounddevice

# === Local Whisper (GPU-Accelerated Transcription) ===
faster-whisper==1.2.0
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# === Data Processing ===
pandas==2.2.3
pydantic==2.10.4
pydantic-settings==2.7.1

# === HTTP Client ===
requests==2.32.3
httpx==0.28.1

# === Security & Authentication ===
passlib==1.7.4
bcrypt==4.0.1  # Compatible with passlib 1.7.4


# === Utilities ===
typing-extensions==4.12.2
aiofiles==24.1.0

# ===================================================================
# IMPORTANT NOTES:
# ===================================================================
# 
# If you see dependency warnings about these packages, you can IGNORE them:
# - langchain-anthropic (not used in this project)
# - langgraph-prebuilt (not used in this project)
# - langgraph-checkpoint-mongodb (not used - we use motor directly)
# - opencv-python (not used in this project)
#
# These packages are from other projects in your environment and won't
# affect TalentFlow AI functionality.
#
# The versions in this file are tested and work together correctly.
# ===================================================================
